Complite Ynadex Game

+++ Assets/Scripts/DataSave/DataToSave/WalletData.cs
@@ -3,7 +3,7 @@ namespace SaveData
     [System.Serializable]
     public class WalletData
     {
-        public int MoneyAmount;
+        public long MoneyAmount;
 
         public WalletData()
         {
